+++ bika/lims/browser/analyses/view.py
@@ -1216,6 +1216,10 @@ class AnalysesView(BikaListingView):
 
         if self.is_analysis_edition_allowed(analysis_brain):
             item["allow_edit"].extend(["Remarks"])
+        else:
+            # render HTMLified text in readonly mode
+            item["Remarks"] = api.text_to_html(
+                analysis_brain.getRemarks, wrap=None)
 
     def _append_html_element(self, item, element, html, glue="&nbsp;",
                              after=True):